Change latitude and longitude to float with said precision

+++ Response/LocationResponse.cs
@@ -2,8 +2,8 @@
 {
     public class LocationResponse
     {
-        public long Latitude { get; set; }
-        public long Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
         public string Place { get; set; } = null!;
     }
 }
